# docs/swagger.yaml
...
openapi: 3.1.1
info:
  title: Lobby API
  version: 1.0.0
  description: API for managing rooms and lobby information.
paths:
...
  /rooms/{id}:
    get:
      summary: Retrieve details of a specific room
      tags:
        - Room Info
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The MongoDB ID of the room
          schema:
            type: string
      responses:
        200:
          description: Room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        404:
          description: Room not found
        500:
          description: Server Error
    delete:
      summary: Delete a room by ID
      tags:
        - Rooms
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Room deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        404:
          description: Room not found
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # -------------------------
  # DATA SCHEMAS
  # -------------------------
  schemas:
    Player:
      type: object
      properties:
        userId:
          type: string
          example: "6585f9e2b1c2d3e4f5a6b7c0"
        name:
          type: string
          example: "MarioRossi"
        imageUrl:
          type: string
          example: "http://example.com/avatar.jpg"
        isHost:
          type: boolean
          example: false
        isReady:
          type: boolean
          example: false
    Room:
      type: object
      properties:
        _id:
          type: string
          example: "6585f9e2b1c2d3e4f5a6b7c8"
        code:
          type: string
          example: "ABC12"
        name:
          type: string
          example: "Lobby Pro"
        gameMode:
          type: string
          enum: [classic, advanced]
          example: "classic"
        roomCapacity:
          type: integer
          example: 4
        players:
          type: array
          items:
            $ref: '#/components/schemas/Player'
        numberOfPlayers:
          type: integer
          description: Virtual field (calculated as players.length)
          example: 1
...